Fix memory leaks, close #4
I was using persistent values too much. This also brings it more in
line with the spec, you can pass an iterator (made from a different
Set or anything that has a .next(), .done and .value properties) to
the constructor. Also different objects that are deeply equal are no
longer considered the same, so you will see something like this:

    > s = new Set
    { size: 0 }
    > s.add({a:1})
    { size: 1 }
    > s.add({a:1})
    { size: 2 }
    > s.add({a:1})
    { size: 3 }
    > s.add({a:1})
    { size: 4 }
    > s.add(1)
    { size: 5 }
    > s.add(1)
    { size: 5 }
    > s.add(1)
    { size: 5 }
    > a = {a:1}
    { a: 1 }
    > s.add(a)
    { size: 6 }
    > s.add(a)
    { size: 6 }
    > s.add(a)
    { size: 6 }
    > s.has({a:1})
    false
    > s.has(a)
    true
    > s.has(1)
    true
    > s.add('a string')
    { size: 7 }
    > s.add('a string')
    { size: 7 }
    > s.add('a string')
    { size: 7 }
    > s.has('a string')
    true

I think this sucks, but it's what the spec says. As a consequence, the
hashing function should be faster and I'm bumping the major version, it
may not be compatible with code already out there.